Remove ActiveRidesVC
